clone: Added Clone/BalanceAfterClone option
The current problem with cloning is that, a clone is always local, this
means the diskful placement will always be the same as of the clone source.
Which leads to uneven usage of node storage.
Now there is a property option to add an additional diskful resource to
the just cloned resource which should pick the least used storage pool
and after the next balance resource task run it will remove the exceeding
resource from the most used storage pool. That should somewhat
achieve a even resource distribution with cloned resources.

Per default this option is disabled and needs to be actively set for now,
we may change the default in future versions.