Node,Del: Fix rare AccessToDeletedNode
In this case the AccessToDeletedData is quite harmless. The controller will not be
able to apply the updates from a satellite to an already deleted node. Since the node
is already deleted, the correct way to handle this case is to simply ignore the delayed
update.
The most harmful in this AccessToDeleteData was the unnecessary ErrorReport.