Drbd: Do not restart DRBD after the first rsc
The previous code set the property that controlls whether or not a local DRBD resource
needs to be restarted even for the very first resource of a resource-definition.
The effect was that the satellite successfully deployed the resource and in the very next
DeviceManagerRun the DRBD resource got restarted. This is completely unnecessary.

Even if we create multiple resources in a loop on the controller side before sending them
out to the satellites, this patch fixes the issue by letting the very first resource set
the min-io property but prevents the restart-DRBD property by checking if it is the very
first resource. That means this patch is build on the the assumption that the
rscMinIoSizeCheck method gets called for every resource individually right after the resource
is created. If that changes for example that resources are first created and then in a
separate loop the rscMinIoSizeCheck method is called for each rsc, this fix no longer works.