Clone: Fix race when starting multiple clones
The issue was that all the clone requests operate during the cloneStart phase on the base resource.
That could mean that one clone sets the base resources into shouldSuspend=true state but before the
satellites can request and apply the change a previous clone can also disable the shouldSuspend again
since it just finished with creating its own snapshot.

This patch fixes the issue by sequencializing all clone requests - at least the cloneStart phase just
as we have already done for snapshots in general. The cloning itself (i.e. a dd if required) is done
asynchronously and in parallel for multiple started clones.